Band-aid database name
There's going to have to be a config switch where I force every single table inside of the "main" database. This is because Django can only connect to hard-coded databases, which makes it impossible to have multiple databases defined by the user. This is going to have to affect every single service in Wywy-Website, from the data entry to the sql-receptionist.